Work-In-Progress-WIP-2020-10-26-122923

+++ src/main/java/com/atkuzmanov/genesys/controllers/DefaultController.java
@@ -3,9 +3,12 @@ package com.atkuzmanov.genesys.controllers;
 import com.atkuzmanov.genesys.dao.TimestampEntity;
 import com.atkuzmanov.genesys.dao.TimestampRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.security.InvalidParameterException;
+import java.sql.Time;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.util.Optional;
@@ -23,47 +26,51 @@ public class DefaultController {
 
     @PostMapping(path = "/addTimestamp")
     public @ResponseBody
-    String addTimestampToDB(@RequestParam String newTimestamp) {
+    ResponseEntity<String> addTimestampToDB(@RequestParam String newTimestamp) {
         TimestampEntity tse = new TimestampEntity();
         if (newTimestamp.isBlank()) {
             newTimestamp = localDateTimeNowFormat_yyyyMMdddHHmmss();
         }
         tse.setTimestampAsString(newTimestamp);
         timestampRepo.save(tse);
-        return "Saved.";
+        return new ResponseEntity<>("Saved.", HttpStatus.OK);
     }
 
     @GetMapping(path = "/all")
     public @ResponseBody
-    Iterable<TimestampEntity> getAllTimestamps() {
+    ResponseEntity<Iterable<TimestampEntity>> getAllTimestamps() {
         // This returns a JSON or XML with all the timestamps.
-        return timestampRepo.findAll();
+        return new ResponseEntity<Iterable<TimestampEntity>>(timestampRepo.findAll(), HttpStatus.OK);
     }
 
-    @PutMapping(path="/updateTimestamp")
+    @PutMapping(path = "/updateTimestamp")
     public @ResponseBody
-    String updateTimestamp(@RequestParam int timestampId) {
+    ResponseEntity<String> updateTimestamp(@RequestParam int timestampId) {
         if (timestampId <= 0) {
             throw new InvalidParameterException("Timestamp id must be a positive integer.");
         }
         Optional<TimestampEntity> optTse = timestampRepo.findById(timestampId);
         if (!optTse.isPresent()) {
-            return "Not found.";
+            return new ResponseEntity<>("Not found.", HttpStatus.NOT_FOUND);
         }
         TimestampEntity tse = optTse.get();
         tse.setTimestampAsString(localDateTimeNowFormat_yyyyMMdddHHmmss());
         timestampRepo.save(tse);
-        return "Updated.";
+        return new ResponseEntity<>("Updated.", HttpStatus.OK);
     }
 
     @DeleteMapping(path = "/delete")
     public @ResponseBody
-    String deleteTimestamp(@RequestParam int timestampId) {
+    ResponseEntity<String> deleteTimestamp(@RequestParam int timestampId) {
         if (timestampId <= 0) {
             throw new InvalidParameterException("Timestamp id must be a positive integer.");
         }
+        Optional<TimestampEntity> optTse = timestampRepo.findById(timestampId);
+        if (!optTse.isPresent()) {
+            return new ResponseEntity<>("Not found.", HttpStatus.NOT_FOUND);
+        }
         timestampRepo.deleteById(timestampId);
-        return "Deleted.";
+        return new ResponseEntity<>(HttpStatus.NO_CONTENT);
     }
 
     private static String localDateTimeNowFormat_yyyyMMdddHHmmss() {
